movement: these durations don't need to be 32-bit ints

+++ movement/movement.c
@@ -8,7 +8,7 @@
 movement_state_t movement_state;
 void * watch_face_contexts[MOVEMENT_NUM_FACES];
 const int32_t movement_le_inactivity_deadlines[8] = {INT_MAX, 3600, 7200, 21600, 43200, 86400, 172800, 604800};
-const int32_t movement_timeout_inactivity_deadlines[4] = {60, 120, 300, 1800};
+const int16_t movement_timeout_inactivity_deadlines[4] = {60, 120, 300, 1800};
 movement_event_t event;
 
 const int16_t movement_timezone_offsets[] = {
